ChangeParamFile {LPJmLmdi}	R Documentation

Change parameters in a parameter file

Description

The function writes values to a parameter file. It requires a 'file.template' in which the positions of the new parameter values are marked with a flag. For example, instead of a parameter of 0.5 for alphaa in a parameter file the flag ALPHAA is written. The function substitutes this flag with the new parameter value in a new file 'file.new'.

Usage

ChangeParamFile(newpar, file.template, file.new, wait = FALSE, 
    ...)

Arguments

newpar
a named vector with new parameter values

file.template
file name of the template parameter with flagged parameters

file.new
file name of the new parameter file

wait
If TRUE wait 1 second to check if file.template exists in order to relax slow file writting.

...
further arguments (currently not used)

Details

The function works only on Unix systems because it is based on 'sed'

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

Examples

# newpar <- c(ALPHAA_BoNE=0.8)
# LPJChangeParamFile(newpar, file.template="pft_template.par", file.new="pft.par")

[Package LPJmLmdi version 1.0 Index]