CombineRescueFiles {LPJmLmdi}	R Documentation

Combine single rescue files into one rescue file

Description

Within OptimizeLPJgenoud, RunLPJ creates rescue file ("_.rescue0.RData") that save the parameter vectors and cost of each individual during optimization. These files allow to create restart files to restart OptimizeLPJgenoud (CreateRestartFromRescue). During OptimizeLPJgenoud many rescue files can be created. The function CombineRescueFiles reads the individual files, combines the rescue objects, saves it in one "rescue.RData" file, and deletes the single files.

Usage

CombineRescueFiles(files.rescue, remove = TRUE)

Arguments

files.rescue
file names

remove
save new rescue file and delete single rescue files?

Details

No details.

Value

The function returns a list of class "rescue", whereby each element corresponds to one individual of the genetic optimization with two entries: 'cost' (cost of the individual) and 'dpar' (parameter scaled relative to the prior parameter).

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

CreateRestartFromRescue, plot.rescue

[Package LPJmLmdi version 1.0 Index]