CostMDS.SSE {LPJmLmdi}	R Documentation

Cost function for multiple data streams based on SSE

Description

The function computes the cost for each grid cell and data stream in 'integrationdata'. Firstly, the cost per data stream and grid cell is computed using the defined 'CostFunction' for each IntegrationDataset. Secondly, the cost is weighted by (1) the dataset-specific weight, (2) the number of observations per grid cell and data streams, and (3) by the grid cell area.

Usage

CostMDS.SSE(integrationdata)

Arguments

integrationdata
object of class 'integrationdata', see IntegrationData

Details

No details.

Value

The function returns a list with the total cost, the cost per grid cell, per data streams, per grid cell and data stream, the error as computed with the defined CostFunction, the number of observations per grid cell and data stream, the weighting factors and grid cell area.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

[Package LPJmLmdi version 1.0 Index]