CreateRestartFromRescue {LPJmLmdi}	R Documentation

Create a *.pro file from binary rescue files to restart optimization

Description

The function creates a *.pro file friom binary 'rescue' files. The *.pro file can be used to restart OptimizeLPJgenoud.

Usage

CreateRestartFromRescue(path.rescue, pop.size)

Arguments

path.rescue
directory where the resuce files from each iteration of the optimization are saved.

pop.size
(estimated) population size of the genetic optimization

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

genoudLPJrescue

[Package LPJmLmdi version 1.0 Index]