EstOptimUse {LPJmLmdi}	R Documentation

Estimate optimal number of jobs given a number of cluster nodes

Usage

EstOptimUse(nodes = 16, wish = 1000)

Arguments

nodes
number of cluster nodes that you want to use

wish
approx. number of elements

Details

No details.

Value

an integer value

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

WriteCLM

[Package LPJmLmdi version 1.0 Index]