FileExistsWait {LPJmLmdi}	R Documentation

Iterative checking and waiting for a file.

Description

The function repeately checks if a file exists and returns TRUE if the file is existing.

Usage

FileExistsWait(file, waitmin = 0, waitinterval = 0.5, waitmax = 2, 
    ...)

Arguments

file
file for which checking and waiting should be applied

waitmin
minimum waiting time (seconds)

waitinterval
interval after which the existence of the file should be checked again (seconds)

waitmax
maximum waiting time (seconds)

...
further arguments (currently not used)

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

Examples

FileExistsWait(system.file("external/rlogo.grd", package="raster"))
FileExistsWait("nofile.txt")

[Package LPJmLmdi version 1.0 Index]