GridProperties {LPJmLmdi}	R Documentation

Derive grid properties from an object of class 'LPJfiles'

Description

The function reads the grid of the input files in 'LPJfiles' and computes the area per grid cell.

Usage

GridProperties(lpjfiles, ...)

Arguments

lpjfiles
list of class 'LPJinput'

...
further arguments (currently not used)

Details

No details.

Value

the function returns a list with 'grid' (raster of grid cells), 'area' (vector of grid cell area) and 'ncell' (number of grid cells)

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

LPJfiles

[Package LPJmLmdi version 1.0 Index]