InfoLPJ {LPJmLmdi}	R Documentation

Information about a LPJmL binary file

Description

The function reads information about a LPJ binary output file.

Usage

InfoLPJ(file.bin = "fpc.bin", file.grid = "grid.bin", file.annual = c("vegc.bin", 
    "litc.bin", "soilc.bin"), size = 4, data.type = numeric(), 
    ...)

Arguments

file.bin
binary LPJ output file

file.grid
binary LPJ grid file

file.annual
one of the binary LPJ output files with annual data

size
the number of bytes per element in the byte stream.

data.type
data type of the file (default=numeric())

...
further arguments (currently not used)

Details

No details.

Value

The function returns a list with information about the LPJ binary file (number of grid cells, number of years, number of bands, spatial extent).

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

ReadLPJ

Examples

# InfoLPJ("vegc.bin")

[Package LPJmLmdi version 1.0 Index]