IntegrationDataset {LPJmLmdi}	R Documentation

Create an object of class 'IntegrationDataset'

Description

The function sets up an object of class 'IntegrationDataset' to define a dataset that should be used in model optimization, including dataset properties and the corresponding model output files. The function also reads the data input files as defined in 'data.val.file' and subsets the input data for the grid cells in 'xy'. One or several 'IntegrationDataset's need to be collected in an object of class IntegrationData which is used in the RunLPJ and OptimizeLPJgenoud functions.

Usage

IntegrationDataset(name, unit = "", data.val.file, data.unc.file, 
    data.time, model.val.file, xy, AggFun = NULL, data.factor = 1, 
    model.factor = 1, cost = TRUE, CostFunction = SSE, weight = 1)

Arguments

name
name of the dataset or variable

unit
unit of the variable (same unit as in LPJmL model output file 'model.val.file')

data.val.file
name of file with the observation values, should be a file that can be read with brick

data.unc.file
name of file with the data uncertainties or a numeric value if the same uncertainty value should be used for all observations

data.time
a vector of class 'Date' with the time period for which the observations are representative. This vector defines also for which time period model results are read to be compared with the data. For example, if data.val.file represents just one value (e.g. long-term mean), the full time period for which this long-term mean is representative needs still to be defined here.

model.val.file
file name of the corresponding model result [e.g. model.val.file="mnpp.bin"] or function without arguments [e.g. model.val.file=function() ReadLPJ("mnpp.bin", start=1901, end=2009, ...) ]. The option to pass a function allows to perform any calculations on LPJ model results or to combine several LPJ model outputs in order to be comparable with observations.

xy
a matrix of grid cells that is used in RunLPJ and OptimizeLPJgenoud. The data in 'data.val.file' and 'data.unc.file' is extracted for these grid cells.

AggFun
aggregation function to aggregate model results to the temporal resolution of the observations, for example AggSumMean for annual sums and mean over annual sums. If NULL no temporal aggregation is done.

data.factor
scaling factor to be applied to the observation data, e.g. for unit conversions

model.factor
scaling factor to be applied to model outputs, e.g. for unit conversions or scaling

cost
Should the data stream be included in the computation of the total cost (TRUE) or not (FALSE). In case of FALSE, evaluation plots are produced for this dataset but the dataset is not considered in the compuation of the total cost and therfore not in optimization.

CostFunction
cost function that should be used for this dataset, default SSE

weight
weighting factor for the dataset in the cost function, cost = CostFunction / number of observations * weight

Details

No details.

Value

The function returns a list of class 'IntegrationDataset'

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

IntegrationData

Examples

# # grid cells for which LPJmL should be run and for which the integration data should be extracted
# xy <- cbind(c(136.75, 137.25, 160.75,168.75), c(45.25, 65.25, 68.75, 63.75))

# # use monthly FAPAR in model-data integration
# fapar <- IntegrationDataset(name="FAPAR", unit="", 
	# data.val.file="GIMMS.FAPAR.1982.2011.nc", 
	# data.unc.file=0.12, 
	# data.time=seq(as.Date("1982-01-01"), as.Date("2011-12-31"), by="month"),
	# model.val.file="mfapar.bin",
	# model.agg=NULL,
	# xy=xy,
	# data.factor=NULL,
	# cost=TRUE,
	# CostFunction=SSE,
	# weight=1)
	
# # use mean annual GPP in model-data integration
# gpp <- IntegrationDataset(name="GPP", unit="gC m-2 yr-1", 
	# data.val.file="MTE.GPP.1982.2011.meanannual.nc", 
	# data.unc.file="MTE.GPPunc.1982.2011.meanannual.nc", 
	# data.time=seq(as.Date("1982-01-01"), as.Date("2011-12-31"), by="month"),
	# model.val.file="mgpp.bin",
	# model.agg=AggSumMean, # sum of each year, mean over all years -> mean annual GPP
	# xy=xy,
	# data.factor=NULL,
	# cost=TRUE,
	# CostFunction=SSE,
	# weight=1)	
	
	

[Package LPJmLmdi version 1.0 Index]