LPJmLmdi-package {LPJmLmdi}	R Documentation

Model-Data Integration for the LPJmL Dynamic Global Vegetation Model

Description

Model-data integration framework for the LPJmL dynamic global vegetation model. Specifically, the package provides functions 1) to optimize LPJmL model parameters using the GENOUD genetic optimization algorithm, 2) to read and write LPJmL input data, and 3) to read LPJmL output files.

Details

The DESCRIPTION file: This package was not yet installed at build time.

Index: This package was not yet installed at build time.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

[Package LPJmLmdi version 1.0 Index]