OptimizeLPJgenoud {LPJmLmdi}	R Documentation

Optimize LPJ using the GENOUD optimizer (genetic optimization using derivatives)

Description

This function performs an optimization of LPJmL model parameters for the specified grid cells unsing the GENOUD genetic optimization algorithm.

Usage

OptimizeLPJgenoud(xy, name, lpjpar, par.optim, lpjfiles, copy.input = TRUE, 
    integrationdata, plot = TRUE, pop.size = 1000, max.generations = 20, 
    wait.generations = 19, BFGSburnin = 18, calc.jacob = FALSE, 
    restart = 0, path.rescue = NULL, restart.jacob = FALSE, nodes = 1, 
    maxAutoRestart = 5, runonly = FALSE, new.spinup4post = TRUE, 
    CostMDS = CostMDS.SSE)

Arguments

xy
matrix of grid cell coordinates to run LPJ

name
name of the experiment (basic file name for all outputs)

lpjpar
data.frame of class LPJpar that define all LPJ parameter values, ranges, and names

par.optim
names of the parameters that should be optimized

lpjfiles
list of class LPJfiles that define all LPJ directories, input files, configuration template files

copy.input
Should LPJ input data be copied to the directory for temporary output? This might speed up computations if the directory is on the same machine where the program runs.

integrationdata
list of integration data and information

plot
plot diagnostic graphics of optimization results?

pop.size
population size, see genoud

max.generations
max number of generations, see genoud

wait.generations
How many generations should genoud wait before returning an optimum, see genoud

BFGSburnin
The number of generations before the L-BFGS-B algorithm is first used, see genoud

calc.jacob
Should the Hessian and Jacobian matrix be computed (yes = TRUE, no = FALSE)?

restart
Where to re-start the optimization? 0 = start at the beginning, 1 = continue with existing genoud optimization, 2 = start after genoud and post-process results.

path.rescue
directory where the resuce files from each iteration of a previous optimization are saved. This is needed if restart > 0.

restart.jacob
Should the Hessian and Jacobian matrix be recomputed if restart > 0 (yes = TRUE, no = FALSE)? Works only if calc.jacob is TRUE.

nodes
use parallel computing? How many nodes to use?

maxAutoRestart
maximum number of automatic restarts of the optimization if an error occurs within genoud()

runonly
run only the model with prior parameters set but don't perform optimization. Produces only results of prior model run.

new.spinup4post
What spinup conditions should be used for the posterior ('posterior-best' and 'posterior-median') model runs? If TRUE, a new spinup is computed based on the optimized parameters. If FALSE, the posterior model runs are started from the spinup conditions of the prior model run (like the runs during optimization).

CostMDS
cost function for multiple data streams to calculate total cost, cost per data stream, and eventually cost per grid cell. See CostMDS.SSE (default) or CostMDS.KGE

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

[Package LPJmLmdi version 1.0 Index]