ReadBIN {LPJmLmdi}	R Documentation

Read simple binary files without header Read a CLM file to a SpatialPixelsDataFrame

Description

The function is used to read for example the soil*.bin and drainclass.bin files

Usage

ReadBIN(file.bin, nbands = 1, size = 1, file.grid = NA, endian.data = "little", 
    endian.grid = "little", data.type = integer(), ...)

Arguments

file.bin
binary file name with extension *.bin

nbands
number of bands per year

size
The number of bytes per element in the byte stream.

file.grid
file name of the corresponding grid file

endian.data
endinaess of the data file

endian.grid
endianess of the grid file

data.type
type of the data

...
Further arguments (currently not used).

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

WriteCLM

Examples

# ReadBIN("soil_new_67420.bin")

[Package LPJmLmdi version 1.0 Index]