ReadCLM {LPJmLmdi}	R Documentation

Read a CLM file to a SpatialPixelsDataFrame

Usage

ReadCLM(file.clm, start = NA, end = NA, start.year = NA, grid = NULL, 
    nbands = NA, size = NA, file.grid = NA, endian.data = NA, 
    endian.grid = "big", data.type = integer(), ...)

Arguments

file.clm
CLM file name with extension *.clm

start
first year to be read

end
last year to be read, reads until last year in case of NA

start.year
first year in the dataset, read from header information in case NA

grid
a matrix of coordinates (lon, lat) if data should be read only for specific cells, if NULL the data for all grid cells is read

nbands
number of bands per year, read from header information in case NA

size
The number of bytes per element in the byte stream.

file.grid
file name of the corresponding grid file

endian.data
endinaess of the data file

endian.grid
endianess of the grid file

data.type
type of the data

...
Further arguments (currently not used).

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

WriteCLM

[Package LPJmLmdi version 1.0 Index]