ReadLPJ {LPJmLmdi}	R Documentation

Read a LPJ binary file

Description

The function reads a binary LPJ output file and returns a SpatialPointsDataFrame

Usage

ReadLPJ(file.bin, file.grid = "grid.bin", sim.start.year = 1901, 
    start = sim.start.year, end = NA, file.annual = c("vegc.bin", 
        "litc.bin", "soilc.bin"), size = 4, data.type = numeric(), 
    endian = "little", ...)

Arguments

file.bin
binary LPJ output file

file.grid
binary LPJ grid file

sim.start.year
first year of the simulation

start
first year to read

end
last year to read, reads until last year if NA

file.annual
one of the binary LPJ output files with annual data

size
the number of bytes per element in the byte stream.

data.type
data type of the file (default=numeric())

endian
endianess of the binary file

...
further arguments (currently not used)

Details

No details.

Value

The function returns a SpatialPointsDataFrame.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

ReadLPJsim

Examples

# ReadLPJ("mgpp.bin", start=1982, end=2011)

[Package LPJmLmdi version 1.0 Index]