ReadLPJ2IntegrationData {LPJmLmdi}	R Documentation

Read LPJ model results into an of class IntegrationData

Description

The function reads for each dataset in IntegrationData the corresponding model output and performs temporal aggregation.

Usage

ReadLPJ2IntegrationData(integrationdata, xy, lpjfiles, ...)

Arguments

integrationdata
object of class IntegrationData

xy
matrix of grid cell coordinates to run LPJ

lpjfiles
list of class LPJfiles that define all LPJ directories, input files, configuration template files

...
further arguments (currently not used)

Details

No details.

Value

The function returns the same list oc class 'IntegrationData' but with added model outputs.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

IntegrationData

[Package LPJmLmdi version 1.0 Index]