ReadLPJ2ts {LPJmLmdi}	R Documentation

Read a LPJ binary file and returns a spatial averaged time series

Description

The function reads LPJ binary output files *.bin, aggregates (mean) the time series over all grid cells and returns the regional-averaged time series

Usage

ReadLPJ2ts(file.bin, sim.start.year = 1901, start = sim.start.year, 
    end = NA, ...)

Arguments

file.bin
binary LPJ output file

sim.start.year
first year of the simulation

start
first year to read

end
last year to read, reads until last year if NA

...
further arguments (currently not used)

Details

No details.

Value

The function returns a time series of class 'ts'.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

ReadLPJsim

Examples

# gpp <- ReadLPJ2ts("mgpp.bin")

[Package LPJmLmdi version 1.0 Index]