ReadLPJinput {LPJmLmdi}	R Documentation

Read and subset CLM files to LPJinput objects

Description

The functions reads a CLM file, selects the data according to the provided grid and returns an object of class LPJinput.

Usage

ReadLPJinput(files, grid = NULL, start = NA, ...)

Arguments

files
character vector of CLM or binary file names

grid
Matrix of grid cells with 2 columns: longitude and latitude (optional). If NULL the data is returned for the grid of the first CLM file. If a grid is provided the data is subesetted for the specified grid cells.

start
first year to read

...
Further arguments to ReadCLM or ReadBIN

Details

No details.

Value

The function returns a list of class "LPJinput".

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

WriteLPJinput

Examples

# lpjinput <- ReadLPJinput("cru_ts_3.20.1901.2011.tmp.clm", grid=cbind(c(136.75, 137.25, 160.75,168.75), c(45.25, 65.25, 68.75, 63.75)))
# str(lpjinput)

[Package LPJmLmdi version 1.0 Index]