ReadOutputvars {LPJmLmdi}	R Documentation

Read 'outputvars.par' to get information about LPJmL output

Description

LPJmL output is defined in par/outputvars.par. This file contains for each variable the id, name, variable name, description, unit, and scale. This file can be used to correctly read LPJmL output. The function is for example used within ReadLPJsim.

Usage

ReadOutputvars(outputvars.par = NULL, ...)

Arguments

outputvars.par
path and file name to the LPJmL 'outputvars.par' file. If NULL the file is searched 1 level above or below the current working directory.

...
further arguments (currently not used)

Details

No details.

Value

The function returns a time series of class 'ts'.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

ReadLPJsim

Examples

# ReadOutputvars()

[Package LPJmLmdi version 1.0 Index]