Rescue2Df {LPJmLmdi}	R Documentation

Convert a 'rescue' list to a data.frame

Description

The function takes an object of class 'rescue' (see CombineRescueFiles) and converts it to a data.frame including the total cost value (1st column), the parameter values (next columns), and the log-likelihood, Akaike's Information Criterion (AIC) and AIC differences (last columns). If 'lpjpar' it not specified the function returns just the scaled parameters (e.g. dpar = par / prior) otherwise it returns the parameters in the original units (e.g. par = dpar * prior).

Usage

Rescue2Df(rescue.l, lpjpar = NULL, ...)

Arguments

rescue.l
a list of class "rescue", see CombineRescueFiles

lpjpar
a list of class "LPJpar" (see LPJpar) to convert the scaled parameters in rescue.l back to the original units (optional)

...
further arguments (currently not used)

Details

No details.

Value

The function returns a data.frame.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

CombineRescueFiles

Examples

# files <- c(list.files(pattern="rescue.RData", recursive=TRUE), list.files(pattern="rescue0.RData", recursive=TRUE))
# rescue.l <- CombineRescueFiles(files, remove=FALSE)
# optim.df <- Rescue2Df(rescue.l)
# summary(optim.df)

[Package LPJmLmdi version 1.0 Index]