Rescue2LPJpar {LPJmLmdi}	R Documentation

Add information from a 'rescue' list to an 'LPJpar' object

Description

The function takes an object of class 'rescue' (see CombineRescueFiles) (alternatively a 'data.frame' as created with Rescue2Df) and a 'LPJpar' (see LPJpar) object. Then it extracts the best parameter set, the median of the best parameter sets (defined based on dAIC <= 2), and various uncertainty measures and adds them to the 'LPJpar' object.

Usage

Rescue2LPJpar(rescue.l, lpjpar, ...)

Arguments

rescue.l
a list of class "rescue" (CombineRescueFiles) or alternatively a data.frame as created with Rescue2Df.

lpjpar
a list of class "LPJpar" (LPJpar)

...
further arguments (currently not used)

Details

No details.

Value

The function returns the provided 'LPJpar' object with the following additional slots:

best Best parameter set

best.median median of best parameter sets (based on all parameter sets with dAIC <= 2)

uncertainty.iqr uncertainty of parameters as the inter-quartile range of the best parameter sets

uncertainty.iqr95 uncertainty of parameters as the central 95% inter-quantile range (0.975-0.025) of the best parameter sets

uncertainty.005.min lower parameter uncertainty estimate based on the minimum parameter value from all parameter sets for which the cost <= quantile(cost, 0.05)

uncertainty.005.max upper parameter uncertainty estimate based on the maximum parameter value from all parameter sets for which the cost <= quantile(cost, 0.05)

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

CombineRescueFiles

Examples

files <- c(list.files(pattern="rescue.RData", recursive=TRUE), list.files(pattern="rescue0.RData", recursive=TRUE))
rescue.l <- CombineRescueFiles(files, remove=FALSE)
lpjpar2 <- Rescue2LPJpar(rescue.l, lpjpar)
str(lpjpar2)
plot(lpjpar2, "ALPHAA", "uncertainty.iqr95")
plot(lpjpar2, "TMIN_BASE", "uncertainty.iqr95")

[Package LPJmLmdi version 1.0 Index]