RunLPJ {LPJmLmdi}	R Documentation

Run LPJmL from R and get results

Description

This function calls LPJmL, reads the results of the model run, computes the cost based on the data sets in IntegrationData and the defined cost function (in CostMDS), and returns the simulation results.

Usage

RunLPJ(dpar, lpjpar, which.par.opt, lpjfiles, path = NULL, integrationdata, 
    xy, newcell = FALSE, name = "LPJmL", plot = FALSE, getresult = FALSE, 
    clean = 1, clean.path = FALSE, CostMDS = CostMDS.SSE, nkeep = 400)

Arguments

dpar
vector of scaling factors for each parameter in 'which.par.opt': parameter = dpar * prior (e.g. if dpar is 1, prior parameters are used in the model run). Optimization is performed on these scaling factors

lpjpar
data.frame of class LPJpar that define LPJ parameter values, ranges, and names

which.par.opt
integer vector that indicates which parameters in lpjpar should be optimized

lpjfiles
list of class LPJfiles that define all LPJ directories, input files, configuration template files

path
path for output files of actual model run

integrationdata
list of of class IntegrationData

xy
matrix of grid cell coordinates to run LPJ

newcell
calculate new cell and new spinup?

name
name of the LPJ run, basic name for all outputs

plot
plot results? see plot.IntegrationData

getresult
If TRUE, all model results are returned in a LPJsim object and model results are saved. If FALSE, only the cost function value is returned.

clean
clean results and temporay configuration and parameter files? 0 = keep everthing; 1 = delete parameter files, conf files and outputs; 2 = clean additionally input files, soil code files and restart

clean.path
Delete output directory 'path' in case it already exists before the model run?

CostMDS
cost function for multiple data streams

nkeep
number of result files to keep. If more are existing, the ones with highest costs will be deleted

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

[Package LPJmLmdi version 1.0 Index]