StartingValues {LPJmLmdi}	R Documentation

Get starting values for genoud from *.pro file

Description

The function extracts the best individuals trhat occured during a genoud optimization from a *.pro file. These best individuals can be used as starting values if a optimization is restarted. This function is called within OptimizeLPJgenoud is a restart is performed.

Usage

StartingValues(file.optresult, pop.size = NULL, ...)

Arguments

file.optresult
genoud *.pro file with optimization results

pop.size
population size

...
further arguments (not used)

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

OptimizeLPJgenoud

[Package LPJmLmdi version 1.0 Index]