WriteCLM {LPJmLmdi}	R Documentation

Write a CLM file from SpatialPointsDataFrame

Description

The function writes CLM files from a SpatialPointsDataFrame or SpatialPixelsDataFrame. The LPJmL program cru2clm needs to be installed.

Usage

WriteCLM(data.sp, file.clm, start, nbands, size = 2, scale = 1, 
    na.replace = -9999, path.lpj = NULL, res = 0.5, ...)

Arguments

data.sp
SpatialPointsDataFrame or SpatialPixelsDataFrame with data

file.clm
CLM file name with extension *.clm

start
integer. First year in data.

nbands
Number of bands per year.

size
The number of bytes per element in the byte stream.

scale
Scaling factor to be written to the header of the CLM file. The factor will be not applied to the data.

na.replace
integer to replace NA values.

path.lpj
path to LPJ installation

res
spatial resolution of the grid cells

...
Further arguments (currently not used).

Details

No details.

Value

The function returns TRUE if the CLM file was created.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

WriteLPJinput

Examples

# data.sp <- SpatialPointsDataFrame(lpjinput$grid, as.data.frame(data.m))
# WriteCLM(data.sp, file="data.clm", start=1901, nbands=12, size=2)	

[Package LPJmLmdi version 1.0 Index]