WriteLPJpar {LPJmLmdi}	R Documentation

Writes an object of class 'LPJpar' as parameter file or table.

Description

The function takes a 'LPJpar' object and writes 1) LPJ parameter files and 2) write *.txt files with parameter values in a table format.

Usage

WriteLPJpar(x, file = "LPJpar", pft.par = NULL, param.par = NULL, 
    param.only = TRUE, ...)

Arguments

x
object of class 'LPJpar'

file
basic file name for all output files, e.g. name of the optimization experiment

pft.par
template file for PFT-specific parameters (create a template from pft.par). If NULL, parameter files will be not written but only parameter tables.

param.par
template file for global parameters (create a template from param.par). If NULL, parameter files will be not written but only parameter tables.

param.only
write only parameters to table (TRUE) or also parameter prior ranges (FALSE)?

...
further arguments for CheckLPJpar

Details

No details.

Value

The function returns a data.frame with an overview of the written files

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

LPJpar, CheckLPJpar

[Package LPJmLmdi version 1.0 Index]