plot.IntegrationData {LPJmLmdi}	R Documentation

Plot an object of class IntegrationData

Description

The function plots an object of class IntegrationData, i.e. it produces a time series plots, scatterplots and a boxplot for the observations and LPJmL model outputs in IntegrationData.

Usage

## S3 method for class 'IntegrationData'
plot(x, ds = 1:length(x), CostMDS = CostMDS.SSE, 
    ...)

Arguments

x
object of class IntegrationData

ds
Which data sets in x should be plotted (integer)

CostMDS
cost function for multiple data streams

...
further arguments (currently not used)

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

IntegrationData

[Package LPJmLmdi version 1.0 Index]