plot.LPJsim {LPJmLmdi}	R Documentation

Plots a LPJsim object

Description

The function plots a LPJsim object: monthly, annual time series or map of grid cells

Usage

## S3 method for class 'LPJsim'
plot(x, what = "annual", start = NA, end = NA, omit0 = TRUE, 
    AggFun = AggMeanNULL, ...)

Arguments

x
an object of class 'LPJsim'

what
What type of plot should be created? 'annual' for yearly time series, 'monthly' for monthly time series, 'daily' for daily time series, and 'grid' for a map of grid cells

start
first year for time series plot

end
last year for time series plot

omit0
omit variables from plotting that are only 0?

AggFun
aggregation function to aggregate results to the temporal resolution as selected in 'what', for example AggMeanNULL for monthly or annual means, AggSumNULL for monthly or annual sums.

...
further arguments (currently not used)

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

ReadLPJ2ts, ReadLPJsim

Examples

# setwd(path.mylpjresult)
# sim <- ReadLPJ2ts(start=1982, end=2011)
# plot(sim, what="annual")

[Package LPJmLmdi version 1.0 Index]