plot.rescue {LPJmLmdi}	R Documentation

plot an object of class "rescue" / monitor OptimizeLPJgenoud

Description

The function plots the cost per data set for all individuals of the genetic optimization from an object of class "rescue". This function can be used to monitor the development of the optimization within OptimizeLPJgenoud. Therefor read the rescue files from your optimization with "rescue.l <- CombineRescueFiles(list.files(pattern=".RData"), remove=FALSE)" and call "plot(rescue.l)".

Usage

## S3 method for class 'rescue'
plot(x, ylim = NULL, xlim = NULL, ylab = "Cost", xlab = "Individuals of genetic optimization", 
    only.cost = FALSE, ...)

Arguments

x
a list of class "rescue", see CombineRescueFiles

ylim
limits of the y-axis of the plot

xlim
limits of the x-axis of the plot

ylab
label for the y axis

xlab
label for the x axis

only.cost
plot all integration datasets (TRUE) or only these ones with cost=TRUE

...
further arguments for plot

Details

No details.

Author(s)

Matthias Forkel <mforkel@bgc-jena.mpg.de> [aut, cre]

References

No reference.

See Also

CombineRescueFiles

[Package LPJmLmdi version 1.0 Index]